style(#781): clear rust-1.94 clippy backlog in adapter/application crates
Root-cause fixes only, no #[allow]; behaviour identical, tests stay green.

- doc_lazy_continuation / doc list indentation (chain wiring, i18n, application
  dispatchers) -> indent continuation lines under their list item or insert a
  blank /// line where a new paragraph followed a list.
- field_reassign_with_default -> struct literal with ..Default::default()
  (selection_highlight, adapter-midi/application tests).
- needless_update (adapter-gui build_device_settings tests): ..default() was
  load-bearing on Linux (3 cfg(target_os="linux") fields); replaced with the
  cfg-gated fields at their default values (realtime/rt_priority/nperiods =
  true/70/3) so it compiles on both and Linux input is byte-identical.
- manual_async_fn (adapter-mcp ServerHandler) -> async fn; redundant_locals
  (compact_chain_param_handlers chain_id); ptr_arg &PathBuf -> &Path
  (project_ops); redundant clone -> slice::from_ref; type_complexity -> type
  aliases; assertions_on_constants -> const { assert!(..) }; let_underscore_future
  (MIDI/bridge submit) -> named _rx binding (the command is enqueued inside
  submit; the reply receiver is intentionally dropped).

Result: `cargo clippy --workspace --all-targets` is warning-clean. Full test
suite unchanged (the one failing io_bindings_ui headless-Slint interaction test
fails identically on the base branch and is unrelated to this cleanup).

+++ crates/adapter-gui/src/io_bindings_ui_tests.rs
@@ -76,7 +76,7 @@ fn binding_named(id: &str, name: &str) -> IoBindingModel {
 fn nth_abs_y(w: &ProjectSettingsWindow, id: &str, n: usize) -> Option<f32> {
     i_slint_backend_testing::ElementHandle::find_by_element_id(w, id)
         .nth(n)
-        .map(|el| el.absolute_position().y as f32)
+        .map(|el| el.absolute_position().y)
 }
 
 #[test]
